requirements.txt
# Python standard libraries (nessuna installazione necessaria)
# os, re, random, warnings già inclusi

# === CORE DEPENDENCIES ===
numpy>=1.24.0,<2.0
pandas>=2.0.0
pyyaml>=6.0

# === NLP LIBRARIES ===
nltk>=3.8
spacy>=3.6.0
gensim>=4.3.0

# === MACHINE LEARNING ===
scikit-learn>=1.3.0
torch>=2.0.0
sentence-transformers>=2.2.0

# === DEEP LEARNING (for Sentiment Classifier) ===
transformers>=4.35.0

# === TEXT PROCESSING ===
emoji>=2.8.0

# === HTTP CLIENT (for Ollama) ===
requests>=2.31.0

# === PROGRESS BARS ===
tqdm>=4.65.0

# === VISUALIZATION (for notebooks) ===
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# === OPTIONAL: Acceleration GPU ===
# Uncomment if you have an NVIDIA GPU
# torch-cuda>=2.0.0

# === DEVELOPMENT TOOLS (optional) ===
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
